pyproject.toml
[tool.poetry]
name = "arbisepolia-scan"
version = "0.1.0"
description = ""
authors = ["oriojas <oscarriojas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
